Refactor the dataset generator
