+++ prompt2model/dataset_generator/prompt_based.py
@@ -8,7 +8,6 @@ import math
 import random
 from collections import Counter, defaultdict
 from dataclasses import dataclass
-from pathlib import Path
 
 import nest_asyncio
 import openai
@@ -36,6 +35,14 @@ class Example:
 
     input_col: str
     output_col: str
+
+    def __eq__(self, other) -> bool:
+        """Example equality."""
+        return self.input_col == other.input_col and self.output_col == other.output_col
+
+    def __lt__(self, other) -> bool:
+        """Example less than."""
+        return self.input_col < other.input_col or self.output_col < other.output_col
 
 
 class PromptBasedDatasetGenerator(DatasetGenerator):
@@ -74,7 +81,6 @@ class PromptBasedDatasetGenerator(DatasetGenerator):
             requests_per_minute: The maximum number of requests per minute.
             filter_duplicated_examples: If True, filters duplicated examples,
                 using the most-frequent output for each input.
-            cache_root: The root directory for caching generated examples.
 
         Raises:
             ValueError: If the 'max_api_calls' value is not greater than 0.
@@ -117,13 +123,13 @@ class PromptBasedDatasetGenerator(DatasetGenerator):
         self.responses_per_request = responses_per_request
         self.requests_per_minute = requests_per_minute
         self.filter_duplicated_examples = filter_duplicated_examples
-        self.cache_root = Path(cache_root)
 
     def construct_prompt(
         self,
         instruction: str,
         few_shot_example_string: str,
         generated_examples: list[Example],
+        context_cutoff: int = 3500,
     ) -> str:
         """Generates a prompt string.
 
@@ -133,53 +139,33 @@ class PromptBasedDatasetGenerator(DatasetGenerator):
         generation, it defaults to using the few_shot_example_string.
 
         The function uses different prompt templates based on the number of selected
-        examples from the generated dataset. If the total length of the prompt exceeds
-        3500 tokens, repeat the prompt generation process to generate a shorter one.
+        examples from the generated dataset.
 
         Args:
             instruction: The natural language instruction for the prompt.
             few_shot_example_string: A string representing the few-shot examples
                 parsed from the user's prompt, which quality is higher than the
-                genrated examples.
+                generated examples.
             generated_examples: A list of currently generated examples.
+            context_cutoff: If the total length of the prompt exceeds this value,
+                repeat the prompt generation process to generate a shorter one.
 
         Returns:
             The generated prompt string.
         """
-        # The random_example_string is a string, which contains several random
-        # few-shot examples as demonstrations for the DatasetGenerator. If
-        # generated_examples is empty, then the random_example_string
-        # is the few-shot examples parsed from the user's prompt.
         while True:
+            # Choose a few examples to add to the prompt if examples exist
             if len(generated_examples) == 0:
                 low_quality_example_string = "N/A\n"
-                # Create default low_quality_example_string if generated_examples
-                # is empty. few_shot_example_string is the high-quality few-shot
-                # examples parsed from the user's prompt. But if user does not
-                # provideany examples in the input prompt, the few_shot_example_string
-                # will be "N/A"/""/None.
                 random_selected_generated_example_num = 0
-                # random_selected_generated_example_num is the number of selected
-                # random examples from generated_examples that will be used to
-                # create the low_quality_example_string. If generated_examples
-                # is empty, then random_selected_generated_example_num is 0.
             else:
-                # If generated_examples is not empty, low_quality_example_string
-                # is sveral random generated examples from generated_examples.
-
                 low_quality_example_string = ""
                 random_selected_generated_example_num = random.randint(
                     1, min(len(generated_examples), 10)
                 )
-                # random_selected_generated_example_num is the number of selected
-                # random examples from generated_examples that will
-                # be concatenated to create low_quality_example_string.
                 random_examples = random.sample(
                     generated_examples, random_selected_generated_example_num
                 )
-                # If generated_examples is not empty, then choose several
-                # random examples from generated_examples to construct
-                # new low_quality_example_string.
                 for example in random_examples:
                     low_quality_example_string += (
                         f'input="{example.input_col}"\noutput="{example.output_col}"\n'
@@ -198,10 +184,7 @@ class PromptBasedDatasetGenerator(DatasetGenerator):
                 high_quality_example_string=few_shot_example_string,
                 template_type=template_type,
             )
-            # The max content length of gpt-3.5-turbo is 4097, so if the
-            # generated prompt is longer than 3500, then the prompt
-            # should be regenerated.
-            if count_tokens_from_string(prompt) < 3500:
+            if count_tokens_from_string(prompt) < context_cutoff:
                 return prompt
             else:
                 orginal_input_string = (
@@ -209,69 +192,17 @@ class PromptBasedDatasetGenerator(DatasetGenerator):
                     if few_shot_example_string
                     else instruction
                 )
-                if count_tokens_from_string(orginal_input_string) > 3500:
+                if count_tokens_from_string(orginal_input_string) > context_cutoff:
                     logger.warning(
-                        "The original input prompt is too long. Consider writing a shorter prompt."  # noqa E501
+                        "The original input prompt is too long. "
+                        "Consider writing a shorter prompt."
                     )
                 continue
 
-    def construct_input_output_map(
+    def apply_multi_vote_filtering(
         self,
         generated_examples: list[Example],
-    ) -> dict[str, Counter]:
-        """Constructs a dictionary mapping inputs to `Counter` objects of outputs.
-
-        Args:
-            generated_examples: A list of currently generated examples.
-
-        Ideally, each input should have a unique output (one-to-one mapping).
-        However, language models may occasionally generate different outputs
-        for identical inputs. For instance, given the input “What is the biggest
-        city in China?”, it might produce different but correct outputs such as
-        “Shanghai” and “The biggest city in China is Shanghai”. At other times,
-        it may produce incorrect variations. For the input “What is the Chemical
-        symbol of gold?”, the outputs might be “Au”, “Au”, and “AU”, where the
-        last one is wrong due to capital letters.
-
-        To address this, PromptBasedDatasetGenerator uses a two-step multi-vote
-        filtering mechanism. This function represents the first step, creating a
-        dictionary to map inputs to a `Counter` of their outputs.
-
-        The function iterates over all the examples, building a dictionary where
-        inputs serve as keys and `Counter` objects as values. The `Counter`
-        tracks the frequency of each output for a specific input.
-
-        For example:
-        input: ["apple", "banana", "apple", "orange", "apple"]
-        output: ["A", "B", "A", "O", "D"]
-
-        Then input_output_map value is:
-        {
-            "apple": Counter({"A": 2, "D": 1}),
-            "banana": Counter({"B": 1}),
-            "orange": Counter({"O": 1})
-        }
-        """
-        input_output_map: dict[str, Counter] = defaultdict(Counter)
-
-        # Iterate through the examples and construct the mapping.
-        for example in generated_examples:
-            input_str = example.input_col
-            output_str = example.output_col
-
-            # Increment the count of the output for the specific input.
-            input_output_map[input_str][output_str] += 1
-
-        # Ensure that the generated_examples list is not empty
-        # and the map is constructed correctly.
-        if len(generated_examples) != 0 and input_output_map is None:
-            raise ValueError("input_output_map is not correctly constructed.")
-
-        return input_output_map
-
-    def apply_multi_vote_to_construct_generated_dataset(
-        self, input_output_map: dict[str, Counter]
-    ) -> Dataset:
+    ) -> list[Example]:
         """Multi-vote to construct generated_dataset from input_output_map.
 
         Args:
@@ -290,35 +221,20 @@ class PromptBasedDatasetGenerator(DatasetGenerator):
         highest frequency, it selects the one that comes first in lexicographical
         (alphabetical) order.
 
-        Example:
-        Suppose input_output_map is:
-        {
-            "apple": Counter({"A": 2, "D": 2}),
-            "banana": Counter({"B": 2, "C": 1}),
-            "orange": Counter({"O": 1})
-        }
-
-        The function will produce generated_dataset:
-        {
-            "input_col": ["apple", "banana", "orange"],
-            "output_col": ["A", "B", "O"]
-        }
-
-        Note: When generated_examples is empty, both input_output_map
-        and generated_dataset will be empty.
-
         Returns:
             Currently generated dataset with multi-vote filtering applied.
         """
-        # Ensure that multi-vote filtering is enabled.
-        if not self.filter_duplicated_examples:
-            raise ValueError("Multi-vote filtering is not enabled.")
+        filtered_examples = []
 
-        filtered_inputs = []
-        filtered_outputs = []
+        input_output_map: dict[str, Counter] = defaultdict(Counter)
+
+        for ex in generated_examples:
+            input_output_map[ex.input_col][ex.output_col] += 1
+
+        if len(generated_examples) != 0 and input_output_map is None:
+            raise ValueError("input_output_map is not correctly constructed.")
 
         for input_str, output_counter in input_output_map.items():
-            # Find the most frequent output count.
             most_common_count = output_counter.most_common(1)[0][1]
 
             # Get all the outputs that have the most common count.
@@ -335,73 +251,10 @@ class PromptBasedDatasetGenerator(DatasetGenerator):
             most_frequent_outputs.sort(key=len)
             final_output = most_frequent_outputs[0]
 
-            filtered_inputs.append(input_str)
-            filtered_outputs.append(final_output)
+            filtered_examples.append(Example(input_str, final_output))
+        return filtered_examples
 
-        # Note that when `generated_examples` is empty,
-        # `input_output_map` is None, and `generated_dataset`
-        # will also be empty.
-        generated_dataset = Dataset.from_dict(
-            {"input_col": filtered_inputs, "output_col": filtered_outputs}
-        )
-        return generated_dataset
-
-    def create_all_examples_dataset_and_generated_dataset(
-        self, generated_examples: list[Example]
-    ) -> Dataset:
-        """Converts generated_examples into generated_dataset.
-
-        Args:
-            generated_examples: A list of currently generated examples.
-
-        Depending on the value of self.filter_duplicated_examples, the function either
-        constructs a mapping for input-output pairs followed by multi-vote filtering
-        to create a Dataset or directly converts the generated examples into a Dataset.
-
-        The function also verifies the presence of data in the input-output map
-        and the generated dataset if there are any generated examples and
-        self.filter_duplicated_examples is True.
-
-        Lastly, the function stores all generated examples, irrespective of the value
-        of self.filter_duplicated_examples, into a Dataset on the disk.
-
-        Returns:
-            A dataset of all the generated examples and the currently generated
-            dataset. If filter_duplicated_examples is True, multi-vote filtering is
-            performed. Else, the generated examples are directly converted into
-            a Dataset.
-        """
-        # Convert all generated examples into a Dataset.
-        all_generated_examples_dataset = Dataset.from_dict(
-            {
-                "input_col": [example.input_col for example in generated_examples],
-                "output_col": [example.output_col for example in generated_examples],
-            }
-        )
-
-        if self.filter_duplicated_examples:
-            # When filtering duplicated examples is
-            # enabled, perform multi-vote filtering.
-            input_output_map = self.construct_input_output_map(generated_examples)
-            generated_dataset = self.apply_multi_vote_to_construct_generated_dataset(
-                input_output_map
-            )
-
-            if len(generated_examples) != 0 and input_output_map is None:
-                raise ValueError("The input-output map is not correctly constructed.")
-        else:
-            # When filtering duplicated examples is not enabled,
-            # use all_generated_examples_dataset directly.
-            generated_dataset = all_generated_examples_dataset
-
-        if len(generated_examples) != 0 and len(generated_dataset) == 0:
-            raise ValueError("The generated dataset is not correctly constructed.")
-
-        return all_generated_examples_dataset, generated_dataset
-
-    def compute_batch_size(
-        self, expected_num_examples: int, generated_dataset: Dataset
-    ) -> int:
+    def compute_batch_size(self, num_examples: int, generated_dataset_size: int) -> int:
         """Computes the batch size for API calls in a batch.
 
         The batch size is determined based on the remaining number of examples to be
@@ -409,10 +262,8 @@ class PromptBasedDatasetGenerator(DatasetGenerator):
         the maximum limit of API calls if it is set.
 
         Args:
-            expected_num_examples: The total number of examples expected to be
-            generated for the current dataset split. Note that if max_api_calls is not
-            set, the actual number of generated examples can be slightly higher due
-            to each API call returning `responses_per_request` examples.
+            num_examples: The total number of examples expected to be
+                generated for the current dataset split.
             generated_dataset: Currently generated dataset.
 
         Returns:
@@ -427,10 +278,7 @@ class PromptBasedDatasetGenerator(DatasetGenerator):
         batch_size = min(
             self.max_batch_size,
             math.ceil(
-                (
-                    (expected_num_examples - len(generated_dataset))
-                    / self.responses_per_request
-                )
+                ((num_examples - generated_dataset_size) / self.responses_per_request)
             ),
             max_api_calls,
         )
@@ -439,15 +287,15 @@ class PromptBasedDatasetGenerator(DatasetGenerator):
             raise ValueError("Batch size must be greater than 0.")
         return batch_size
 
-    def extract_responses(
+    def extract_and_append_responses(
         self, completions: list[openai.Completion], generated_examples: list[Example]
-    ) -> list[Example]:
+    ) -> None:
         """Extracts the generated sample and annotation from an API response.
 
         Args:
             completions: A list of Completion objects returned by the API.
-            Each API call returns a number of completion objects equivalent to
-            `responses_per_request`. The default `responses_per_request` = 5.
+                Each API call returns a number of completion objects equivalent to
+                `responses_per_request`. The default `responses_per_request` = 5.
             generated_examples: Currently generated examples of DatasetGenerator.
 
         This function iterates through the provided completions, attempting to
@@ -459,44 +307,6 @@ class PromptBasedDatasetGenerator(DatasetGenerator):
         with `input_col` and `output_col` fields, representing the generated
         example and label strings respectively. The `example` is then added
         to generated_examples.
-
-        Note: The function process `batch_size * responses_per_request`
-        responses at a time.
-
-        Example:
-            Given a list of two completion objects: [completion_1, completion_2],
-            where:
-            completion_1.choices = [
-                {"message": {"content": '{"input": "1", "output": "a"}'}},
-                {"message": {"content": '{"input": "1", "output": "b"}'}},
-                {"message": {"content": '{"input": "1", "output": "a"}'}},
-            ]
-            completion_2.choices = [
-                {"message": {"content": '{"input": "1", "output": "c"}'}},
-                {"message": {"content": '{"input": "2", "output": "a"}'}},
-                {"message": {"content": '{"input": "2", "output": "b"}'}},
-            ]
-
-            The function will create 'example' namedtuples:
-            Example(input_col="1", output_col="a")
-            Example(input_col="1", output_col="b")
-            Example(input_col="1", output_col="a")
-            Example(input_col="1", output_col="c")
-            Example(input_col="2", output_col="a")
-            Example(input_col="2", output_col="b")
-
-            It will then append them to generated_examples.
-
-        Returns:
-            A list of `Example` objects.
-                Each API call will return `responses_per_request` completion objects.
-                If the response is a valid JSON object, create a namedtuple called
-                `example` and append it to generated_examples. `example` consists
-                of `input_col` and`output_col`, where:
-                - input_col is the generated example string extracted from the response.
-                - output_col is the generated label string extracted from the response.
-                If the response is not a valid JSON object, discard it.
-                There is responses_per_request * len(completions) responses at a time.
         """
         for completion in completions:
             try:
@@ -532,12 +342,11 @@ class PromptBasedDatasetGenerator(DatasetGenerator):
                     f"Error happened when parsing API completion: {completion}"
                 )
                 continue
-        return generated_examples
 
     async def generate_responses(
         self,
         chat_api: APIAgent,
-        generated_dataset: Dataset,
+        generated_dataset_size: int,
         expected_num_examples: int,
         prompts: list[str],
     ) -> list[openai.Completion]:
@@ -563,9 +372,12 @@ class PromptBasedDatasetGenerator(DatasetGenerator):
         """
         # Calculate the dynamic temperature based
         # on the size of the generated dataset
-        dynamic_temperature = (self.max_temperature - self.initial_temperature) * len(
-            generated_dataset
-        ) / expected_num_examples + self.initial_temperature
+        dynamic_temperature = (
+            (self.max_temperature - self.initial_temperature)
+            * generated_dataset_size
+            / expected_num_examples
+            + self.initial_temperature
+        )
 
         # Ensure the dynamic temperature is within the range [0, 2.0]
         clipped_temperature = max(0.0, min(2.0, dynamic_temperature))
@@ -580,12 +392,12 @@ class PromptBasedDatasetGenerator(DatasetGenerator):
     def generate_dataset_split(
         self,
         prompt_spec: PromptSpec,
-        expected_num_examples: int,
-        split: DatasetSplit,
+        num_examples: int,
+        split: DatasetSplit = DatasetSplit.TRAIN,
     ) -> Dataset:
-        """Generates a dataset split using GPT-3.5.
+        """Generates a dataset split using API-based LMs.
 
-        This method iteratively makes API calls to GPT-3.5 to generate a dataset split.
+        This method iteratively makes API calls to generate a dataset split.
         Each API call yields a batch of responses. From these responses, new examples
         are extracted and added to 'generated_examples'. The process continues
         until the desired number of examples is reached, or the maximum limit on API
@@ -594,97 +406,63 @@ class PromptBasedDatasetGenerator(DatasetGenerator):
 
         Args:
             prompt_spec: PromptParser to be used for generating examples.
-            expected_num_examples: The number of examples expected to be
-                generated. If the maximum limit on API calls is not set, the actual
-                number of generated examples can be slightly higher due to each
-                API call returning `responses_per_request` examples.
-            split: The dataset split (e.g., train, validation, test) for which the
-                examples are being generated.
+            num_examples: The number of examples to be generated.
 
         Returns:
             The generated dataset split.
         """
-        # Refresh the relevant data structures for the new split.
-        self.cache_root.mkdir(parents=True, exist_ok=True)
-        examples_cache_path = Path(
-            self.cache_root / f"generated_examples_{split.value}"
-        )
-        dataset_cache_path = Path(self.cache_root / f"generated_dataset_{split.value}")
+        all_generated_examples: list[Example] = []
+        generated_examples: list[Example] = []
 
-        if examples_cache_path.exists():
-            # If cache exists, load generated examples from disk.
-            logger.info(f"Loading cache from {str(examples_cache_path)}.")
-            all_generated_examples_dataset = Dataset.load_from_disk(examples_cache_path)
-            generated_examples = [
-                Example(input_col=ex["input_col"], output_col=ex["output_col"])
-                for ex in all_generated_examples_dataset
-            ]
-        else:
-            # Initialize data structures for a new split.
-            generated_examples = []
-
-        pbar = tqdm(total=expected_num_examples, desc="Generating examples")
+        pbar = tqdm(total=num_examples, desc="Generating examples")
         chat_api = APIAgent()
 
-        while True:
-            # Each API call will return `responses_per_request` completion
-            # objects. The upper bound of the length of the generated dataset
-            # is expected_num_examples + responses_per_request.
-            try:
-                # Convert the generated examples into a
-                # Dataset and update the progress bar.
-                (
-                    all_generated_examples_dataset,
-                    generated_dataset,
-                ) = self.create_all_examples_dataset_and_generated_dataset(
-                    generated_examples
+        while len(generated_examples) < num_examples:
+            if self.max_api_calls and self.api_call_counter >= self.max_api_calls:
+                logger.warning("Maximum number of API calls reached.")
+                break
+
+            batch_size = self.compute_batch_size(num_examples, len(generated_examples))
+            self.api_call_counter += batch_size
+
+            # Generate prompts for the batch call.
+            prompts = [
+                self.construct_prompt(
+                    instruction=prompt_spec.instruction,
+                    few_shot_example_string=prompt_spec.examples,
+                    generated_examples=generated_examples,
                 )
-                all_generated_examples_dataset.save_to_disk(examples_cache_path)
-                generated_dataset.save_to_disk(dataset_cache_path)
-                pbar.update(len(generated_dataset))
+                for _ in range(batch_size)
+            ]
 
-                if self.max_api_calls and self.api_call_counter >= self.max_api_calls:
-                    logger.warning("Maximum number of API calls reached.")
-                    break
-                elif len(generated_dataset) >= expected_num_examples:
-                    break
-                else:
-                    # Compute the batch size for the next API call.
-                    batch_size = self.compute_batch_size(
-                        expected_num_examples, generated_dataset
+            try:
+                loop = asyncio.get_event_loop()
+                responses = loop.run_until_complete(
+                    self.generate_responses(
+                        chat_api=chat_api,
+                        generated_dataset_size=len(generated_examples),
+                        expected_num_examples=num_examples,
+                        prompts=prompts,
                     )
-                    self.api_call_counter += batch_size
-
-                    # Generate prompts for the batch call.
-                    prompts = [
-                        self.construct_prompt(
-                            instruction=prompt_spec.instruction,
-                            few_shot_example_string=prompt_spec.examples,
-                            generated_examples=generated_examples
-                            if not self.filter_duplicated_examples
-                            else [
-                                Example(each["input_col"], each["output_col"])
-                                for each in generated_dataset
-                            ],
-                        )
-                        for _ in range(batch_size)
-                    ]
-
-                    loop = asyncio.get_event_loop()
-                    responses = loop.run_until_complete(
-                        self.generate_responses(
-                            chat_api=chat_api,
-                            generated_dataset=generated_dataset,
-                            expected_num_examples=expected_num_examples,
-                            prompts=prompts,
-                        )
-                    )
-
-                    # Extract the responses and add new examples to the dataset.
-                    generated_examples = self.extract_responses(
-                        responses, generated_examples
-                    )
+                )
             except API_ERRORS as e:
-                # Handle API errors and adjust the API call counter.
-                self.api_call_counter = handle_api_error(e, self.api_call_counter)
-        return generated_dataset
+                handle_api_error(e)
+
+            # Extract the responses and add new examples to the dataset.
+            self.extract_and_append_responses(responses, all_generated_examples)
+            generated_examples = (
+                self.apply_multi_vote_filtering(all_generated_examples)
+                if self.filter_duplicated_examples
+                else all_generated_examples
+            )
+            if len(generated_examples) >= num_examples:
+                generated_examples = generated_examples[:num_examples]
+
+            pbar.update(len(generated_examples))
+
+        return Dataset.from_dict(
+            {
+                "input_col": [ex.input_col for ex in generated_examples],
+                "output_col": [ex.output_col for ex in generated_examples],
+            }
+        )
